app/src/main/java/mega/privacy/android/app/lollipop/megachat/ChatActivityLollipop.java
app/src/main/java/mega/privacy/android/app/lollipop/megachat/chatAdapters/MegaChatLollipopAdapter.java

+++ app/src/main/java/mega/privacy/android/app/UploadService.java
@@ -73,8 +73,6 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 	private boolean isForeground = false;
 	private boolean canceled;
 
-	private boolean isQRFile = false;
-
 	MegaApplication app;
 	MegaApiAndroid megaApi;
 	MegaChatApiAndroid megaChatApi;
@@ -197,7 +195,7 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 		log("onHandleIntent");
 
 		final File file = new File(intent.getStringExtra(EXTRA_FILEPATH));
-		isQRFile = intent.getBooleanExtra("qrfile", false);
+
 		if(file!=null){
 			log("File to manage: "+file.getAbsolutePath());
 		}
@@ -528,8 +526,12 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 					inProgress = inProgress + currentTransfer.getTransferredBytes();
 				}
 			}
-			long inProgressTemp = inProgress *100;
-			progressPercent = inProgressTemp/total;
+
+			long inProgressTemp = 0;
+			if(total>0){
+				inProgressTemp = inProgress *100;
+				progressPercent = inProgressTemp/total;
+			}
 
 			String message = "";
 			if (inProgress == 0){
@@ -855,15 +857,17 @@ public class UploadService extends Service implements MegaTransferListenerInterf
 					}
 				}
 
-				if (getApplicationContext().getExternalCacheDir() != null && !isQRFile) {
+				String qrFileName = megaApi.getMyEmail() + "QRcode.jpg";
+
+				if (getApplicationContext().getExternalCacheDir() != null) {
 					File localFile = new File(getApplicationContext().getExternalCacheDir(), transfer.getFileName());
-					if (localFile.exists()) {
+					if (localFile.exists() && !localFile.getName().equals(qrFileName)) {
 						log("Delete file!: " + localFile.getAbsolutePath());
 						localFile.delete();
 					}
-				} else if (!isQRFile){
+				} else {
 					File localFile = new File(getApplicationContext().getCacheDir(), transfer.getFileName());
-					if (localFile.exists()) {
+					if (localFile.exists() && !localFile.getName().equals(qrFileName)) {
 						log("Delete file!: " + localFile.getAbsolutePath());
 						localFile.delete();
 					}
